7월 업데이트

+++ presentation/screens/Main/Mypage.tsx
@@ -1,5 +1,5 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import React, { useEffect, useState, useRef, useCallback } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import {
   Image,
   StyleSheet,
@@ -9,19 +9,23 @@ import {
   TouchableOpacity,
   Modal,
   Pressable,
+  Dimensions,
+  Alert,
 } from 'react-native';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import { Avatar, Switch } from 'react-native-elements';
+import {
+  SafeAreaView,
+  useSafeAreaInsets,
+} from 'react-native-safe-area-context';
+import { Avatar } from 'react-native-elements';
 import WebView from 'react-native-webview';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import Toast from 'react-native-easy-toast'
+import Toast from 'react-native-easy-toast';
 import { useIsFocused } from '@react-navigation/native';
 import { privateAPI } from '../../../infrastructures/api/remote/base';
 import { NavigationStackParams } from '../../../infrastructures/types/NavigationStackParams';
 import ButtonBack from '../../components/ButtonBack';
 import COLORS from '../../styles/colors';
 import { commonStyles } from '../../styles/common';
-
 
 type MyPageScreenProp =
   | NativeStackScreenProps<NavigationStackParams, 'MyPage'> & {
@@ -31,6 +35,8 @@ type MyPageScreenProp =
 const mypageLine = require('../../../assets/mypage-line.png');
 const icCloseTerms = require('../../../assets/ic-close-terms.png');
 const IcArrowRight = require('../../../assets/ic-arrow-right.png');
+
+const windowWidth = Dimensions.get('window').width;
 
 const styles = StyleSheet.create({
   headerContainer: {
@@ -128,7 +134,7 @@ const styles = StyleSheet.create({
   },
   toastPopup: {
     backgroundColor: '#161D24',
-    width: 380, // figma: 345
+    width: windowWidth - 30,
     opacity: 0.6,
     paddingVertical: 15,
     paddingHorizontal: 25,
@@ -149,6 +155,7 @@ const styles = StyleSheet.create({
 
 function MyPageScreen({ navigation, resetUserState }: MyPageScreenProp) {
   const isFocused = useIsFocused();
+  const insets = useSafeAreaInsets();
   const [isEditable, setIsEditable] = useState(true);
   const [isSetAlarm, setIsSetAlarm] = useState(true);
 
@@ -189,18 +196,31 @@ function MyPageScreen({ navigation, resetUserState }: MyPageScreenProp) {
     navigation.navigate('SignUpMain');
   };
 
-  const withdrawl = () => {
-    privateAPI.patch({ url: 'api/v1/users/auth/status' }).then((response) => {
-      if (response.code === 'SUCCESS') {
-        AsyncStorage.clear();
-        resetUserState();
-        navigation.navigate('SignUpMain');
-      }
-    });
+  const withdrawal = () => {
+    Alert.alert('경고', '회원 정보가 삭제됩니다. 탈퇴하시겠습니까?', [
+      {
+        text: '취소',
+        style: 'cancel',
+      },
+      {
+        text: '탈퇴하기',
+        onPress: () => {
+          privateAPI
+            .patch({ url: 'api/v1/users/auth/status' })
+            .then((response) => {
+              if (response.code === 'SUCCESS') {
+                AsyncStorage.clear();
+                resetUserState();
+                navigation.navigate('SignUpMain');
+              }
+            });
+        },
+      },
+    ]);
   };
 
   const onMembershipPressed = () => {
-    toastRef.current?.show('아직 준비중이에요! 7월에 만나요-!');
+    toastRef.current?.show('멤버십은 아직 준비 중이에요!');
   };
 
   return (
@@ -210,9 +230,7 @@ function MyPageScreen({ navigation, resetUserState }: MyPageScreenProp) {
         <ButtonBack onPress={() => navigation.goBack()} />
         <Text style={styles.headerTitle}>마이페이지</Text>
         <TouchableOpacity
-          onPress={() =>
-            navigation.navigate('EditProfile', { nickname: nickName })
-          }
+          onPress={() => navigation.navigate('EditProfile')}
           disabled={!isEditable}
         >
           <Text
@@ -283,6 +301,12 @@ function MyPageScreen({ navigation, resetUserState }: MyPageScreenProp) {
             </Pressable>
           </View>
           <Image source={mypageLine} style={styles.lineImage} />
+          <View style={styles.eachSettings}>
+            <Pressable onPress={() => navigation.navigate('PasswordChange')}>
+              <Text style={styles.settingTxt}>비밀번호 변경</Text>
+            </Pressable>
+          </View>
+          <Image source={mypageLine} style={styles.lineImage} />
 
           {/* 로그아웃 */}
           <View style={styles.eachSettings}>
@@ -302,7 +326,7 @@ function MyPageScreen({ navigation, resetUserState }: MyPageScreenProp) {
             </Text>
             <Image source={IcArrowRight} style={styles.membershipArrowIcon} />
           </TouchableOpacity>
-          <Text style={styles.withdrawTxt} onPress={withdrawl}>
+          <Text style={styles.withdrawTxt} onPress={withdrawal}>
             탈퇴하기
           </Text>
         </View>
@@ -312,7 +336,7 @@ function MyPageScreen({ navigation, resetUserState }: MyPageScreenProp) {
         ref={toastRef}
         style={styles.toastPopup}
         position="bottom"
-        positionValue={25}
+        positionValue={insets.bottom + 168}
         fadeInDuration={200}
         fadeOutDuration={1000}
         textStyle={styles.toastTxt}
